RF - Separate direction finder from model fit.
